+++ test/mavsdk_tests/integration_test_runner/logger_helper.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import re
 import sys
 import os
